+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Adding `JsonbListView` with `isSqlNull(int index)` method to check if a JSONB array has SQL or JSON null value.
 - Fix double-timezone encoding bug.
+- Fix statement scope bug.
 
 ## 3.5.10